[Transport] One vocabulary for who is told; map ids stop being narrowed
The deck/shore couriers had multiplied along the wrong axis: every new kind of
thing that had to cross got its own function, so the vessel had one, the crew had
one, the map icons had one, and a passenger had none -- which is why boarding was
invisible from the pier. The subject never varied. The AUDIENCE did.

Map::Audience { Here, ShoreWatch, Voyage, AnchorAll } names it once, and
CollectAudience/Relay answer it once. AnnounceAboard and AnnounceZoneMapTracked
fold into Relay; CollectRelayAudience now asks in the same words.

The three pairs that were one function with a direction become one each:
AnnounceVessel/RetractVessel -> RelayVessel(give), where the direction is only
which end of the packet the hull goes; AppendCrew{Create,Destroy}Blocks ->
AppendCrewBlocks(give). SetExternalObservers had a single caller inside the class
that owns the invariant, so it is gone and GatherObservers writes the member.

Seventeen couriers down to fourteen, and all of them now say "who" the same way.

Persistence: MapPersistentState and its three subclasses took the map id as
uint16 and were handed MapEntry::ID, a uint32. Lossless only because 3.3.5a stops
below 1000. A vessel's hull is a map like any other and goes through here.

+++ src/game/WorldHandlers/MapPersistentStateMgr.h
@@ -67,7 +67,7 @@ class MapPersistentState
 {
         friend class MapPersistentStateManager;
     protected:
-        MapPersistentState(uint16 MapId, uint32 InstanceId, Difficulty difficulty);
+        MapPersistentState(uint32 MapId, uint32 InstanceId, Difficulty difficulty);
 
     public:
 
@@ -140,6 +140,12 @@ class MapPersistentState
         typedef std::unordered_map<uint32, time_t> RespawnTimes;
 
         uint32 m_instanceid;
+
+        /// uint32, and every constructor down the chain takes it as one. It used to be
+        /// narrowed to uint16 at each of them from MapEntry::ID, which is a uint32 --
+        /// lossless only because 3.3.5a happens to stop below 1000. A vessel's hull is a
+        /// map like any other and goes through here; nothing about it should depend on the
+        /// highest map id Blizzard shipped.
         uint32 m_mapid;
         Difficulty m_difficulty;
         Map* m_usedByMap;                                   // NULL if map not loaded, non-NULL lock MapPersistentState from unload
@@ -163,7 +169,7 @@ class WorldPersistentState : public MapPersistentState
            - any new non-instanceable map created
            - respawn data loading for non-instanceable map
         */
-        explicit WorldPersistentState(uint16 MapId) : MapPersistentState(MapId, 0, REGULAR_DIFFICULTY) {}
+        explicit WorldPersistentState(uint32 MapId) : MapPersistentState(MapId, 0, REGULAR_DIFFICULTY) {}
 
         ~WorldPersistentState() {}
 
@@ -191,7 +197,7 @@ class DungeonPersistentState : public MapPersistentState
            - any new instance is being generated
            - the first time a player bound to InstanceId logs in
            - when a group bound to the instance is loaded */
-        DungeonPersistentState(uint16 MapId, uint32 InstanceId, Difficulty difficulty, time_t resetTime, bool canReset, uint32 completedEncountersMask);
+        DungeonPersistentState(uint32 MapId, uint32 InstanceId, Difficulty difficulty, time_t resetTime, bool canReset, uint32 completedEncountersMask);
 
         ~DungeonPersistentState();
 
@@ -265,7 +271,7 @@ class BattleGroundPersistentState : public MapPersistentState
         /* Created either when:
            - any new BG/arena is being generated
         */
-        BattleGroundPersistentState(uint16 MapId, uint32 InstanceId, Difficulty difficulty)
+        BattleGroundPersistentState(uint32 MapId, uint32 InstanceId, Difficulty difficulty)
             : MapPersistentState(MapId, InstanceId, difficulty) {}
 
         ~BattleGroundPersistentState() {}
@@ -299,7 +305,7 @@ struct DungeonResetEvent
 {
     ResetEventType type   : 8;                              // if RESET_EVENT_NORMAL_DUNGEON then InstanceID == 0 and applied to all instances for pair (map,diff)
     Difficulty difficulty : 8;                              // used with mapid used as for select reset for global cooldown instances (instamceid==0 for event)
-    uint16 mapid;                                           // used with mapid used as for select reset for global cooldown instances (instanceid==0 for event)
+    uint32 mapid;                                           // used with mapid used as for select reset for global cooldown instances (instanceid==0 for event)
     uint32 instanceId;                                      // used for select reset for normal dungeons
 
     DungeonResetEvent() : type(RESET_EVENT_NORMAL_DUNGEON), difficulty(DUNGEON_DIFFICULTY_NORMAL), mapid(0), instanceId(0) {}
